arreglo bugs menores

+++ src/main/java/com/marcorh96/springboot/datajpa/app/springbootdatajpa/models/services/UploadFileServiceImpl.java
@@ -21,7 +21,7 @@ public class UploadFileServiceImpl implements IUploadFileService {
 
     private final Logger log = LoggerFactory.getLogger(getClass());
 
-    private final static String UPLOADS_FOLDER = "spring-boot-data-jpa/uploads";
+    private final static String UPLOADS_FOLDER = "uploads";
 
     @Override
     public Resource load(String filename) throws MalformedURLException {
@@ -46,13 +46,16 @@ public class UploadFileServiceImpl implements IUploadFileService {
 
     @Override
     public boolean delete(String filename) {
-        Path rooPath = this.getPath(filename);
-            File archivo = rooPath.toFile();
+        if (filename != null && filename.length() > 0) {
+            Path rootPath = getPath(filename);
+            File archivo = rootPath.toFile();
+
             if (archivo.exists() && archivo.canRead()) {
                 if (archivo.delete()) {
                     return true;
                 }
             }
+        }
         return false;
     }
 
@@ -63,13 +66,13 @@ public class UploadFileServiceImpl implements IUploadFileService {
     @Override
     public void deleteAll() {
         FileSystemUtils.deleteRecursively(Paths.get(UPLOADS_FOLDER).toFile());
-        
+
     }
 
     @Override
     public void init() throws IOException {
         Files.createDirectory(Paths.get(UPLOADS_FOLDER));
-        
+
     }
 
 }
